% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_annotations.R
\name{clean_annotations}
\alias{clean_annotations}
\title{Clean and compile ROV annotation files}
\usage{
clean_annotations()
}
\value{
a dataframe
}
\description{
Cleans ROV annotations data from the ASPIRE campaign
}
\details{
Cool blurb about NOAA Exploration
}
\examples{

ASPIRE_dat <- clean_annotations()

}
\seealso{
\code{\link{https://github.com/shgroves/NOAA.ASPIRE}}
}
