% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_EX1806.R
\name{clean_EX1806}
\alias{clean_EX1806}
\title{Clean EX1806 annotations found in the comments column}
\usage{
clean_EX1806()
}
\value{
a dataframe of EX1806 dives
}
\description{
Removes annotation from the Comments column to the Taxonomy column and pulls Taxon Path from WoRMS
}
\details{
Cool blurb about NOAA Exploration
}
\examples{

EX1806_dat <- clean_EX1806()

}
\seealso{
\code{\link{https://github.com/shgroves/NOAA.ASPIRE}}
}
