% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_SHIP_CTD.R
\name{compile_SHIP_CTD}
\alias{compile_SHIP_CTD}
\title{Compile CTD Rosette}
\usage{
compile_SHIP_CTD(expeditions, path, return_dataframe = "NULL")
}
\arguments{
\item{path}{A string indicating the filepath to the .hex files}

\item{return_dataframe}{A logical value (TRUE or FALSE) indicating whether a dataframe should be returned.}
}
\value{
This function returns a dataframe of CTD casts
}
\description{
Compiles all CTD rosette data from cnv files into a single dataframe
}
\details{
NOAA Ocean Exploration collects oceanographic data and water samples by request during mapping and ROV
expeditions on \emph{Okeanos Explorer} with a CTD rosette system. The CTD rosette contains conductivity,
temperature, depth, oxygen, salinity, oxidation reduction potential, turbidity, and fluorometer sensors.
CTD rosette data are archived as .hex and .cnv files, which are produced using Seabird software.
}
\examples{
compile_SHIP_CTD(path = "C:/Users/sarah.groves/Documents/Data/CTD/", campaign = "ASPIRE", return_dataframe = T)
}
\seealso{
\code{\link{NOAA.ASPIRE::getROVCTD, NOAA.ASPIRE::getCTDRosette}}
}
