% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hex_to_cnv.R
\name{convert_hex_to_cnv}
\alias{convert_hex_to_cnv}
\title{Convert .hex to .cnv}
\usage{
convert_hex_to_cnv(path1, path2 = "NULL", path3 = "NULL", path4 = "NULL")
}
\arguments{
\item{path}{A string indicating the filepath to the .hex files}
}
\value{
This function returns a .cnv file to the specified path
}
\description{
Converts .hex files to .cnv files
}
\details{
NOAA Ocean Exploration collects oceanographic data and water samples by request during mapping and ROV
expeditions on \emph{Okeanos Explorer} with a CTD rosette system. The CTD rosette contains conductivity,
temperature, depth, oxygen, salinity, oxidation reduction potential, turbidity, and fluorometer sensors.
CTD rosette data are archived as .hex and .cnv files, which are produced using Seabird software.
}
\examples{
# You must specify at least one path
path1 <- "C:/Users/sarah.groves/Documents/Data/CTD/ASPIRE_EX1905L1/SHIPCTD"
path2 <- "C:/Users/sarah.groves/Documents/Data/CTD/ASPIRE_EX1906/SHIPCTD"
}
\seealso{
\code{\link{NOAA.ASPIRE::getROVCTD, NOAA.ASPIRE::getCTDRosette}}
}
