% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_SHIP_CTD_files.R
\name{count_SHIP_CTD_files}
\alias{count_SHIP_CTD_files}
\title{Count ship CTD Rosette files}
\usage{
count_SHIP_CTD_files(expeditions, data_inventory, path)
}
\arguments{
\item{expeditions}{A list indicating the expeditions of interest (must be from the ASPIRE campaign where CTD rosette data were collected; must be '"1805"', '"1810"','"1812"', '"1903l1"','"1905l1"', '"1906"','"2101"','"2102"','"2107"')}

\item{data_inventory}{The ASPIRE campaign data inventory}

\item{path}{A filepath to where the CTD files are located.}
}
\value{
a dataframe with CTD file and data structure and counts
}
\description{
Counts ship CTD files and filepath structure
}
\details{
Cool blurb about NOAA Exploration
}
\examples{

First run:
ASPIRE_dat <- googlesheets4::read_sheet('https://docs.google.com/spreadsheets/d/19jRXe7SeDQRjv-guH2po5xTHqfzfeicDzZ74cb3OnNY/edit?pli=1#gid=0',
.name_repair = "universal")

count_SHIP_CTD_files(expeditions = expeditions,
                    path = "C:/Users/sarah.groves/Documents/Data/CTD/",
                    data_inventory = ASPIRE_dat)

}
\seealso{
\code{\link{https://github.com/shgroves/NOAA.ASPIRE}}
}
