% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NOAA_WOA_data.R
\name{get_NOAA_WOA_data}
\alias{get_NOAA_WOA_data}
\alias{url_parser}
\title{Function to obtain NOAA World Ocean Atlas dataset}
\usage{
get_NOAA_WOA_data(var, spat_res, av_period, cache = FALSE)

url_parser(var, spat_res, av_period, cache = FALSE)
}
\arguments{
\item{var}{The chemical or physical variable of interest (possible choices:
\code{"temperature"}, \code{"conductivity"}, \code{"oxygen"}, \code{"mixed layer depth"},
\code{"salinity"}, \code{"density"}).}

\item{spat_res}{Spatial resolution, for temperature and salinity, either 1 or .25 degree grid-cells (numeric).}

\item{av_period}{Temporal resolution, either \code{"annual"} or month (Will pull the month the cast was made from the data).
Defaults to annual if month is not available in the data.}

\item{cache}{Caching the extracted NOAA file in the package's \code{extdata}
directory (default = \code{FALSE}). Size of individual files is around 12 Mb.}
}
\value{
\code{\link[stars:st_as_stars]{stars}} object
}
\description{
Retrieves data from the NOAA World Ocean Atlas. Adapted from the function "get_NOAA" from the oceanexplorer package by Martin Schobben
}
\details{
Calls: url_parser.R, read_NetCDF.R

Called by: extract_NOAA_WOA.R, NOAA Explore QAQC_CTD Rosette.Rmd

NOAA Exploration Science and Technology Data Analysis team: Groves, Egan
Last update: Oct. 2023

Functions to retrieve data from the
\href{https://www.ncei.noaa.gov/products/world-ocean-atlas}{NOAA World Ocean Atlas}
. Data is an 3D array (longitude, latitude, and depth) and is loaded as a
\code{\link[stars:st_as_stars]{stars}} object. Check \code{\link{NOAA_data}} for available
variables, respective units and their citations. The function can automatically
cache the extracted files (default: \code{cache = FALSE}). The cached file will
then reside in the package's \code{extdata} directory.
}
\examples{

# path to NOAA server or local data source
url_parser("density", 1, "annual")

if (curl::has_internet() && interactive()) {

# retrieve NOAA data
get_NOAA_WOA_data("oxygen", 1, "annual")

}
}
\seealso{
\href{https://r-spatial.github.io/stars/articles/stars1.html}{Introduction to the stars package}, code{\link{https://github.com/shgroves/NOAA.ASPIRE}}
}
