% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_campaign_metrics.R
\name{make_campaign_metrics}
\alias{make_campaign_metrics}
\title{Make campaign metrics}
\usage{
make_campaign_metrics(link)
}
\arguments{
\item{link}{A string indicating the the link to the Google sheet}
}
\value{
a dataframe with campaign metrics summary
}
\description{
Computes campaign metrics from Google sheet
}
\details{
Cool blurb about NOAA Exploration
}
\examples{
# You must specify the link
link <- "https://docs.google.com/spreadsheets/d/19jRXe7SeDQRjv-guH2po5xTHqfzfeicDzZ74cb3OnNY/edit?pli=1#gid=0"
}
\seealso{
\code{\link{https://github.com/shgroves/NOAA.ASPIRE}}
}
