% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_ROVCTD.R
\name{pull_ROVCTD}
\alias{pull_ROVCTD}
\title{Pull ROV CTD data from data landing page based on specified parameters}
\usage{
pull_ROVCTD(expedition, dive, start_time, end_time, path)
}
\arguments{
\item{expedition}{A list indicating the expeditions of interest (must be from the ASPIRE campaign where an ROV dive was conducted, do not include "EX" before the number; acceptable values are: "1806" "1811" "1903" "1904" "1905" "1907" "2103" "2104" "2107" "2201" "2205" "2206")}

\item{dive}{The dive of interest, must be in a two-digit format in " ", for example, "01", "02", "03", etc.}

\item{start_time}{The start time of the cast of interest, must be in the XXX format. Default value is the beginning of the descent. These can be pulled from the ROV Text file in the ancillary data download.}

\item{end_time}{The end time of the cast of interest, must be in the XXX format. Default is when the ROV hits bottom. These can be pulled from the ROV Text file in the ancillary data download.}

\item{path}{A file path to where you would like your file to be exported to.}
}
\value{
a data frame of an ROV cast
}
\description{
pull an ROV CTD file from the data landing page and clip the dive based on provided expedition, dive, and start and end times.
}
\details{
Cool blurb about NOAA Exploration
}
\seealso{
\code{\link{https://github.com/shgroves/NOAA.ASPIRE}}
}
