% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_NetCDF.R
\name{read_netCDF}
\alias{read_netCDF}
\title{Function to read netcdf file given a URL and attribute}
\usage{
read_netCDF(conn, var)
}
\arguments{
\item{conn}{a URL leading to the netCDF file}

\item{var}{a variable representing the attributes in the NOAA WOA netCDF files}
}
\value{
\code{\link[stars:st_as_stars]{stars}} object
}
\description{
Adapted from the function "read_NOAA" from the oceanexplorer package by Martin Schobben
}
\details{
Called by: get_NOAA_WOA.R

NOAA Exploration Science and Technology Data Analysis team: Groves, Egan
Last update: Oct. 2023

Functions to retrieve data from a given url for a netCDF file.
Data are a 3D array (longitude, latitude, and depth) and is loaded as a
\code{\link[stars:st_as_stars]{stars}} object. Check \code{\link{https://data.nodc.noaa.gov/woa/WOA18/DOC/woa18documentation.pdf}} for available
variables, respective units and their citations. See url_parser function in get_NOAA_WOA_data.R for help with creating a URL from variables.
}
\examples{


if (curl::has_internet() && interactive()) {

# retrieve netCDF and convert to stars object
read_netCDF("https://www.ncei.noaa.gov/thredds-ocean/dodsC/woa23/DATA/temperature/netcdf/decav91C0/1.00/woa23_decav91C0_t08_01.nc",
            "t_an")

}
}
\seealso{
\href{https://r-spatial.github.io/stars/articles/stars1.html}{Introduction to the stars package}, code{\link{https://github.com/shgroves/NOAA.ASPIRE}}
}
